Bug: Problem with XSPI with higher freqs. Important: Reference does use a slow scaled (prescaler = 3) freq to configurate and afer that it bypasses the prescaler and switch the freq up to 200mhz

+++ embedded/STM32N6570DK/FSBL/Src/Drivers/Simple/Inc/simple_xspi.h
@@ -9,8 +9,8 @@ typedef struct {
 	uint8_t chipselect_high_time;
 	uint8_t prescaler;
 	uint8_t chipselect_boundary;
-	uint8_t maxtran_value;
-	uint8_t refresh_cycles;
+	uint16_t maxtran_value;
+	uint16_t refresh_cycles;
 } XSPI_cfg_TypeDef;
 
 typedef struct {
